Sueastside did more work on the environment manager.

+++ src/client/environment/sky/utils/tone_reproductor.cpp
@@ -41,7 +41,7 @@ void ToneReproductor::set_display_adaptation_luminance(float _Lda)
 
 	// Update terms
 	alpha_wa_over_alpha_da = alpha_wa/alpha_da;
-	term2 = pow10((beta_wa-beta_da)/alpha_da) / (M_PI*0.0001f);
+	term2 = pow10((beta_wa-beta_da)/alpha_da) / ((float)M_PI*0.0001f);
 }
 
 // Set the eye adaptation luminance for the world and precompute what can be
@@ -56,7 +56,7 @@ void ToneReproductor::set_world_adaptation_luminance(float _Lwa)
 
 	// Update terms
 	alpha_wa_over_alpha_da = alpha_wa/alpha_da;
-	term2 = pow10((beta_wa-beta_da)/alpha_da) / (M_PI*0.0001f);
+	term2 = pow10((beta_wa-beta_da)/alpha_da) / ((float)M_PI*0.0001f);
 
 }
